refactored spawn daemon to not have circular dependencies
